+++ Abaya_Store.Application/Features/ProductReviews/Handlers/Commands/DeleteProductReviewCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using Abaya_Store.Application.Features.ProductReviews.Requests.Commands;
+﻿using Abaya_Store.Application.Exceptions;
+using Abaya_Store.Application.Features.ProductReviews.Requests.Commands;
 using Abaya_Store.Application.Persistence.Contracts;
 using AutoMapper;
 using MediatR;
@@ -19,6 +20,10 @@ namespace Abaya_Store.Application.Features.ProductReviews.Handlers.Commands
 		public async Task<Unit> Handle(DeleteProductReviewCommand request, CancellationToken cancellationToken)
 		{
 			var review = await _productReviewRepository.GetByIdAsync(request.Id);
+
+			if (review == null)
+				throw new NotFoundException(nameof(review), request.Id);
+
 			await _productReviewRepository.DeleteAsync(review);
 			return Unit.Value;
 		}
