+++ Abaya_Store.Application/Features/ShippingInfos/Handlers/Commands/DeleteShippingInfoCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using Abaya_Store.Application.Features.ShippingInfos.Requests.Commands;
+﻿using Abaya_Store.Application.Exceptions;
+using Abaya_Store.Application.Features.ShippingInfos.Requests.Commands;
 using Abaya_Store.Application.Persistence.Contracts;
 using AutoMapper;
 using MediatR;
@@ -19,6 +20,10 @@ namespace Abaya_Store.Application.Features.ShippingInfos.Handlers.Commands
 		public async Task<Unit> Handle(DeleteShippingInfoCommand request, CancellationToken cancellationToken)
 		{
 			var shippingInfo = await _shippingInfoRepository.GetByIdAsync(request.Id);
+
+			if (shippingInfo == null)
+				throw new NotFoundException(nameof(shippingInfo), request.Id);
+
 			await _shippingInfoRepository.DeleteAsync(shippingInfo);
 			return Unit.Value;
 		}
