+++ Abaya_Store.Application/Features/WishListItems/Handlers/Commands/DeleteWishListItemCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using Abaya_Store.Application.Features.WishListItems.Requests.Commands;
+﻿using Abaya_Store.Application.Exceptions;
+using Abaya_Store.Application.Features.WishListItems.Requests.Commands;
 using Abaya_Store.Application.Persistence.Contracts;
 using AutoMapper;
 using MediatR;
@@ -19,6 +20,10 @@ namespace Abaya_Store.Application.Features.WishListItems.Handlers.Commands
 		public async Task<Unit> Handle(DeleteWishListItemCommand request, CancellationToken cancellationToken)
 		{
 			var item = await _wishListItemRepository.GetByIdAsync(request.Id);
+
+			if (item == null)
+				throw new NotFoundException(nameof(item), request.Id);
+
 			await _wishListItemRepository.DeleteAsync(item);
 			return Unit.Value;
 		}
