+++ Abaya_Store.Application/Features/WishLists/Handlers/Commands/DeleteWishListCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using Abaya_Store.Application.Features.WishLists.Requests.Commands;
+﻿using Abaya_Store.Application.Exceptions;
+using Abaya_Store.Application.Features.WishLists.Requests.Commands;
 using Abaya_Store.Application.Persistence.Contracts;
 using AutoMapper;
 using MediatR;
@@ -19,6 +20,10 @@ namespace Abaya_Store.Application.Features.WishLists.Handlers.Commands
 		public async Task<Unit> Handle(DeleteWishListCommand request, CancellationToken cancellationToken)
 		{
 			var wishList = await _wishListRepository.GetByIdAsync(request.Id);
+
+			if (wishList == null)
+				throw new NotFoundException(nameof(wishList), request.Id);
+
 			await _wishListRepository.DeleteAsync(wishList);
 			return Unit.Value;
 		}
